requirements.txt
alembic==1.8.1
autopep8==1.7.0
click==8.1.3
dominate==2.7.0
Flask==2.2.2
Flask-Bootstrap==3.3.7.1
Flask-Cors==3.0.10
Flask-Login==0.6.2
Flask-Migrate==3.1.0
Flask-SQLAlchemy==3.0.1
Flask-WTF==1.0.1
greenlet==1.1.3.post0
itsdangerous==2.1.2
Jinja2==3.1.2
Mako==1.2.3
MarkupSafe==2.1.1
psycopg2==2.9.4
pycodestyle==2.9.1
python-dotenv==0.21.0
six==1.16.0
SQLAlchemy==1.4.41
toml==0.10.2
visitor==0.1.3
Werkzeug==2.2.2
WTForms==3.0.1
